+++ module/Application/src/Application/Entity/UserWrapper.php
@@ -5,7 +5,7 @@ namespace Application\Entity;
 use Individu\Entity\Db\Individu;
 use Application\Entity\Db\Utilisateur;
 use Exception;
-use UnicaenAuth\Entity\Ldap\People as UnicaenAuthPeople;
+use UnicaenApp\Entity\Ldap\People as UnicaenAppPeople;
 use UnicaenApp\Exception\LogicException;
 use UnicaenApp\Exception\RuntimeException;
 use UnicaenAuth\Entity\Db\AbstractUser;
@@ -17,7 +17,7 @@ use ZfcUser\Entity\UserInterface;
  * Wrapper représentant un utilisateur authentifié, permettant de masquer autant que faire se peut
  * les différences entre les classes d'utilisateurs pouvant être rencontrées dans l'appli :
  * - UnicaenLdapPeople
- * - UnicaenAuthPeople
+ * - UnicaenAppPeople
  * - Utilisateur
  * - ShibUser
  *
@@ -26,7 +26,7 @@ use ZfcUser\Entity\UserInterface;
 class UserWrapper implements UserInterface
 {
     /**
-     * @var UnicaenLdapPeople|UnicaenAuthPeople|Utilisateur|ShibUser
+     * @var UnicaenLdapPeople|UnicaenAppPeople|Utilisateur|ShibUser
      */
     private $userData;
 
@@ -36,7 +36,7 @@ class UserWrapper implements UserInterface
     private $individu;
 
     /**
-     * @param Utilisateur|UnicaenAuthPeople|ShibUser|UnicaenLdapPeople $userData
+     * @param Utilisateur|UnicaenAppPeople|ShibUser|UnicaenLdapPeople $userData
      * @return self
      * @throws \Exception
      */
@@ -44,7 +44,7 @@ class UserWrapper implements UserInterface
     {
         switch (true) {
             case $userData instanceof UnicaenLdapPeople:
-            case $userData instanceof UnicaenAuthPeople:
+            case $userData instanceof UnicaenAppPeople:
             case $userData instanceof ShibUser:
             case $userData instanceof Utilisateur:
                 $this->userData = $userData;
@@ -52,7 +52,7 @@ class UserWrapper implements UserInterface
             default:
                 throw new Exception(
                     "Type de données utilisateurs spécifié inattendu : " .
-                    is_object($userData) ? get_class($userData) : gettype($userData)
+                    (is_object($userData) ? get_class($userData) : gettype($userData))
                 );
         }
 
@@ -125,7 +125,7 @@ class UserWrapper implements UserInterface
     {
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 return $this->userData->getEduPersonPrincipalName();
 
             case $this->userData instanceof ShibUser:
@@ -151,7 +151,7 @@ class UserWrapper implements UserInterface
     {
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 $supannId = $this->userData->getSupannEmpId() ?: $this->userData->getSupannEtuId();
                 break;
 
@@ -187,7 +187,7 @@ class UserWrapper implements UserInterface
 
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 return $this->userData->getSn(true);
 
             case $this->userData instanceof Utilisateur:
@@ -212,7 +212,7 @@ class UserWrapper implements UserInterface
 
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 return $this->userData->getGivenName();
 
             case $this->userData instanceof Utilisateur:
@@ -237,7 +237,7 @@ class UserWrapper implements UserInterface
 
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 return $this->userData->getSupannCivilite();
 
             case $this->userData instanceof Utilisateur:
@@ -264,7 +264,7 @@ class UserWrapper implements UserInterface
     {
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
             case $this->userData instanceof Utilisateur:
             case $this->userData instanceof ShibUser:
                 return $this->userData->getId();
@@ -295,7 +295,7 @@ class UserWrapper implements UserInterface
             case $this->userData instanceof UnicaenLdapPeople:
                 return $this->userData->getSupannAliasLogin();
 
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
             case $this->userData instanceof Utilisateur:
             case $this->userData instanceof ShibUser:
                 return $this->userData->getUsername();
@@ -324,7 +324,7 @@ class UserWrapper implements UserInterface
     {
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 return $this->userData->getMail();
 
             case $this->userData instanceof Utilisateur:
@@ -355,7 +355,7 @@ class UserWrapper implements UserInterface
     {
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 return $this->userData->getNomComplet(true);
 
             case $this->userData instanceof Utilisateur:
@@ -386,7 +386,7 @@ class UserWrapper implements UserInterface
     {
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 return 'ldap';
 
             case $this->userData instanceof Utilisateur:
@@ -419,7 +419,7 @@ class UserWrapper implements UserInterface
     {
         switch (true) {
             case $this->userData instanceof UnicaenLdapPeople:
-            case $this->userData instanceof UnicaenAuthPeople:
+            case $this->userData instanceof UnicaenAppPeople:
                 $parts = ldap_explode_dn($this->userData->getDn(), 1);
                 $isDeactivated = in_array('deactivated', $parts);
                 return $isDeactivated ? 0 : 1;